semantic-search/requirements.txt
azure-functions==1.17.0
azure-core==1.29.5
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==0.28.1
python-dotenv==1.0.0
numpy==1.24.3
requests==2.31.0
typing-extensions>=4.8.0